+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/automation/configuration/ConfigurationVraNg.java
@@ -17,10 +17,10 @@ package com.vmware.pscoe.iac.artifact.aria.automation.configuration;
 import java.net.URISyntaxException;
 import java.util.Properties;
 
-import org.springframework.util.StringUtils;
 import org.apache.hc.core5.http.HttpHost;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.util.StringUtils;
 
 import com.vmware.pscoe.iac.artifact.configuration.Configuration;
 import com.vmware.pscoe.iac.artifact.configuration.ConfigurationException;