+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/vcd/configuration/ConfigurationVcd.java
@@ -23,9 +23,16 @@ import com.vmware.pscoe.iac.artifact.model.PackageType;
 public final class ConfigurationVcd extends Configuration {
 	// Important - when modify properties refer to comments in @Configuration
 	/**
-	 * vCD Package Import content conflict resolution mode
+	 * VCD Package Import content conflict resolution mode
 	 */
 	public static final String PACKAGE_IMPORT_OVERWRITE_MODE = "packageImportOverwriteMode";
+
+	public static final String USER_AT_DOMAIN_STRING_FORMAT = "%s@%s";
+
+	/**
+	 * param ORGANIZATION_NAME.
+	 */
+	public static final String ORGANIZATION_NAME = "org.name";
 
 	private ConfigurationVcd(Properties props) {
 		super(PackageType.VCDNG, props);
@@ -33,6 +40,13 @@ public final class ConfigurationVcd extends Configuration {
 
 	public String getPackageImportOverwriteMode() {
 		return this.properties.getProperty(PACKAGE_IMPORT_OVERWRITE_MODE, "SKIP,OVERWRITE");
+	}
+
+	/**
+	 * @return String
+	 */
+	public String getOrgName() {
+		return this.properties.getProperty(ORGANIZATION_NAME);
 	}
 
 	public static ConfigurationVcd fromProperties(Properties props) throws ConfigurationException {